# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tabme API",
    "version": "1.0.0",
    "description": "The backend API for the Tabme Chrome Extension. It provides services for user authentication, data storage, and synchronization of spaces, folders, bookmarks, and sticky notes across devices."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Space": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID"
          },
          "title": {
            "type": "string",
            "description": "Space title"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering spaces"
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Folder"
            }
          },
          "widgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Widget"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Folder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Folder ID"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID this folder belongs to"
          },
          "title": {
            "type": "string",
            "description": "Folder title"
          },
          "color": {
            "type": "string",
            "description": "Folder color"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering folders"
          },
          "twoColumn": {
            "type": "boolean",
            "description": "Whether folder displays in two columns"
          },
          "collapsed": {
            "type": "boolean",
            "description": "Whether folder is collapsed"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookmarkItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Bookmark item ID"
          },
          "folderId": {
            "type": "string",
            "format": "uuid",
            "description": "Folder ID this item belongs to"
          },
          "title": {
            "type": "string",
            "description": "Bookmark title"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Bookmark URL"
          },
          "favIconUrl": {
            "type": "string",
            "format": "uri",
            "description": "Favicon URL"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering items"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Widget": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Widget ID"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID this widget belongs to"
          },
          "pos": {
            "type": "object",
            "properties": {
              "x": {
                "type": "number"
              },
              "y": {
                "type": "number"
              },
              "w": {
                "type": "number"
              },
              "h": {
                "type": "number"
              },
              "zIndex": {
                "type": "number"
              }
            }
          },
          "content": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "sticky-note"
                ]
              },
              "text": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SyncChange": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE",
              "RESTORE"
            ]
          },
          "entityType": {
            "type": "string",
            "description": "Type of entity (space, folder, item, widget)"
          },
          "entityId": {
            "type": "string",
            "description": "ID of the entity"
          },
          "data": {
            "type": "object",
            "description": "Entity data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SyncLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "operation": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE",
              "RESTORE"
            ]
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/sync/full": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Get full sync data",
        "description": "Retrieve all user data for complete synchronization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full sync data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "spaces": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Space"
                      }
                    },
                    "folders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Folder"
                      }
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BookmarkItem"
                      }
                    },
                    "widgets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Widget"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User data not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sync/changes": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Get sync changes",
        "description": "Retrieve incremental sync changes since last sync version",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lastSyncVersion",
            "schema": {
              "type": "integer"
            },
            "description": "Last sync version number"
          },
          {
            "in": "query",
            "name": "entityTypes",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of entity types to sync"
          }
        ],
        "responses": {
          "200": {
            "description": "Sync changes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "changes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SyncChange"
                      }
                    },
                    "currentVersion": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sync/apply": {
      "post": {
        "tags": [
          "Sync"
        ],
        "summary": "Apply sync changes",
        "description": "Apply a batch of sync changes to the server",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "changes"
                ],
                "properties": {
                  "changes": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SyncChange"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sync changes applied successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or some changes failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sync/log": {
      "post": {
        "tags": [
          "Sync"
        ],
        "summary": "Create sync log entry",
        "description": "Create a new sync log entry for tracking changes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "operation",
                  "entityType",
                  "entityId"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "CREATE",
                      "UPDATE",
                      "DELETE",
                      "RESTORE"
                    ]
                  },
                  "entityType": {
                    "type": "string"
                  },
                  "entityId": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sync log created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncLog"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sync/stats": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Get user sync statistics",
        "description": "Retrieve sync statistics for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sync statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lastSyncAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "syncCount": {
                      "type": "integer"
                    },
                    "dataVersion": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sync/cleanup": {
      "delete": {
        "tags": [
          "Sync"
        ],
        "summary": "Cleanup old sync logs",
        "description": "Remove old sync logs (admin endpoint)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "daysToKeep",
            "schema": {
              "type": "integer",
              "default": 30
            },
            "description": "Number of days to keep sync logs"
          }
        ],
        "responses": {
          "200": {
            "description": "Cleanup completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}